add simplify util
