+++ test/cpp/pir/cinn/adt/index_expr_test.cc
@@ -20,6 +20,7 @@
 #include "paddle/cinn/ir/ir_base.h"
 #include "paddle/cinn/ir/ir_mutator.h"
 #include "paddle/cinn/ir/op/ir_operators.h"
+#include "paddle/cinn/optim/simplify_util.h"
 
 namespace cinn {
 namespace common {
@@ -525,13 +526,13 @@ TEST_F(TestIndexExpr, TestCheckPattern) {
   ir::IndexExpr e1 = (S0 * (S1 + S2) + S1 * S2 + S2) / (S4 * S5) * S4 +
                      (S0 * (S1 + S2) + S1 * S2 + S2) % (S4 * S5) / S5;
   std::unordered_map<std::string, ir::IndexExpr> map;
-  EXPECT_TRUE(common::CheckPattern(e, pattern, &map));
+  EXPECT_TRUE(optim::CheckPattern(e, pattern, &map));
   map.clear();
-  EXPECT_FALSE(common::CheckPattern(e, pattern1, &map));
+  EXPECT_FALSE(optim::CheckPattern(e, pattern1, &map));
   map.clear();
-  EXPECT_FALSE(common::CheckPattern(e1, pattern, &map));
+  EXPECT_FALSE(optim::CheckPattern(e1, pattern, &map));
   map.clear();
-  EXPECT_TRUE(common::CheckPattern(e1, pattern1, &map));
+  EXPECT_TRUE(optim::CheckPattern(e1, pattern1, &map));
 }
 }  // namespace common
 }  // namespace cinn
